pyproject.toml
[project]
name = "azure-cost-agent-langgraph"
version = "0.1.0"
description = "Azure cost optimizer — LangGraph implementation with supervisor routing"
requires-python = ">=3.13"
license = "MIT"
dependencies = [
    "langgraph>=1.1.0",
    "langgraph-supervisor>=0.0.30",
    "langchain>=1.2.0,<2.0",
    "pyyaml>=6.0",
    "langchain-openai>=1.1.0",
    "azure-identity>=1.25.0",
    "azure-mgmt-costmanagement>=4.0.0",
    "azure-mgmt-resourcegraph>=8.0.0",
    "azure-mgmt-advisor>=9.0.0",
    "azure-mgmt-monitor>=6.0.0",
    "azure-mgmt-consumption>=10.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "chainlit>=2.10.0",
    "langchain-azure-ai[opentelemetry]>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.9.0",
    "mypy>=1.19.0",
    "types-PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "S", "SIM", "TCH", "RUF", "PTH", "PERF", "LOG"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S106"]

[tool.mypy]
python_version = "3.13"
strict = true
# Exclude third-party boundary modules (chainlit/langchain/langgraph
# lack type stubs, making strict checking impossible in these files).
exclude = "src/(workflow|app|cli|foundry_main|agents/reporter)\\.py$"

[[tool.mypy.overrides]]
module = [
    "langchain_azure_ai.*",
    "langgraph.*",
    "langgraph_supervisor.*",
    "langgraph_prebuilt.*",
    "chainlit.*",
    "langchain_core.*",
    "langchain_openai.*",
    "langchain.*",
    "azure.*",
]
ignore_missing_imports = true


[tool.uv]
prerelease = "allow"
